pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-forge"
version = "0.1.0"
description = "Engineering-grade reusable agent framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.9.0",
  "typer>=0.12.0",
  "openai>=2.24.0",
  "tavily-python>=0.5.0",
  "pydantic-settings>=2.13.1",
  "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
]

[project.scripts]
agent-forge = "agent_forge.apps.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration", "tests/e2e"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
